feat(28): #10 deprecate per-platform build scripts, move hook config
BUILD-01: Delete build-claude.ts (263 lines) and build-cursor.ts (256 lines),
remove build:claude and build:cursor package.json scripts. All output is now
generated exclusively through build-all.ts -> generateAllOutputs().

BUILD-02: Move generateClaudeHooksConfig() from scripts/build-shared.ts to
src/hooks/index.ts, co-locating it with generateCursorHooksConfig(). Fix
index.ts boundary violation (scripts/ -> src/ dependency inversion).

Also: update stale error message in build-all.ts referencing deleted compiler
classes (CLEAN-01 partial).

992 tests pass, 313 build outputs, zero drift.
