+++ package.json
@@ -17,8 +17,6 @@
   "scripts": {
     "build": "bun run --filter '*' build",
     "build:all": "bun run ./scripts/build-all.ts",
-    "build:claude": "bun run ./scripts/build-claude.ts",
-    "build:cursor": "bun run ./scripts/build-cursor.ts",
     "generate:from-cursor": "bun run ./scripts/generate-agents-from-cursor.ts && bun run ./scripts/generate-skills-from-cursor.ts && bun run ./scripts/generate-rules-from-cursor.ts",
     "check:drift": "bun run ./scripts/check-drift.ts",
     "test": "bun test",
